+++ libs/data_processing/include/serializer.hpp
@@ -3,7 +3,7 @@
 #include <string>
 #include <vector>
 
-#include "Block.hpp"
+#include "block.hpp"
 
 class Serializer {
   static std::string ToString(int value);
