+++ libs/network/include/netmessage.hpp
