+++ libs/network/src/netserver.cpp
@@ -1,3 +1,3 @@
-#include "NetServer.hpp"
+#include "netserver.hpp"
 
 bool NetServer::InitializeServer(uint32_t port) {}
